+++ sdk/core/Azure.Core/src/Response.cs
@@ -83,6 +83,14 @@ namespace Azure
         public abstract void Dispose();
 
         /// <summary>
+        /// Indicates whether the message's <see cref="ResponseClassifier"/> considers this
+        /// response an error.
+        /// </summary>
+        public bool IsError { get; internal set; }
+
+        internal HttpMessageSanitizer? Sanitizer { get; set; }
+
+        /// <summary>
         /// Returns header value if the header is stored in the collection. If header has multiple values they are going to be joined with a comma.
         /// </summary>
         /// <param name="name">The header name.</param>