+++ sdk/storage/Azure.Storage.Common/src/Shared/StorageVersionExtensions.cs
@@ -37,7 +37,7 @@ namespace Azure.Storage
         // TODO https://github.com/Azure/azure-sdk-for-net/issues/25140
         public const ServiceVersion LatestVersion =
 #if BlobSDK || QueueSDK || FileSDK || DataLakeSDK || ChangeFeedSDK
-            ServiceVersion.V2020_12_06;
+            ServiceVersion.V2021_02_12;
 #else
             ERROR_STORAGE_SERVICE_NOT_DEFINED;
 #endif