+++ sdk/cognitiveservices/Vision.Face/CHANGELOG.md
@@ -1,7 +1,17 @@
 # Release History
 
-## 2.7.0-preview.2 (Unreleased)
+## 2.7.0-preview.3 (Unreleased)
 
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 2.7.0-preview.2 (2021-11-15)
+- Add support for new detection attribute `qualityForRecognition`
 
 ## 2.7.0-preview.1 (2021-03-25)